requirements.txt
Django>=4.0.4,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
flake8>=4.0.1,<4.1
Pillow>=9.1.0,<9.2